Added inverse FFT function and tests
Description:
This commit introduces an implementation of the inverse Fast Fourier Transform (FFT) function using the iterative FFT function. The inverse FFT function takes a vector of complex numbers as input and returns the inverse FFT of the input signal.

Additionally, this commit includes tests for the new inverse FFT function. The tests compare the output of the custom inverse FFT function with MATLAB's built-in IFFT function and display the results.

Signed-off-by: Noam Rahat <noamrht@gmail.com>